move ProductAttachmentUpdater to Update ns
