+++ src/Adapter/Product/Update/ProductAttachmentUpdater.php
@@ -26,7 +26,7 @@
 
 declare(strict_types=1);
 
-namespace PrestaShop\PrestaShop\Adapter\Product;
+namespace PrestaShop\PrestaShop\Adapter\Product\Update;
 
 use Attachment;
 use PrestaShop\PrestaShop\Adapter\Attachment\AttachmentRepository;
